+++ components/remote-controls/api-backend/src/commonMain/kotlin/com/flipperdevices/ifrmvp/backend/model/UiPresetModel.kt
@@ -4,10 +4,11 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-@SerialName("DeviceCategory")
-class DeviceCategory(
+data class UiPresetModel(
     @SerialName("id")
     val id: Long,
-    @SerialName("meta")
-    val meta: CategoryMeta,
+    @SerialName("infrared_file_id")
+    val infraredFileId: Long,
+    @SerialName("file_name")
+    val fileName: String
 )