+++ components/remote-controls/core-model/src/commonMain/kotlin/com/flipperdevices/ifrmvp/model/buttondata/ChannelButtonData.kt
@@ -8,7 +8,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class ChannelButtonData(
     @SerialName("add_key_id")
-    val addKeyIdentifier: IfrKeyIdentifier,
+    val addKeyIdentifier: IfrKeyIdentifier = IfrKeyIdentifier.Unknown,
     @SerialName("reduce_key_id")
     val reduceKeyIdentifier: IfrKeyIdentifier,
 ) : ButtonData {