+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - [FIX] Crash on app startup with WearOS app
 - [FIX] Expand verify signal bottom sheet only after signal is dispatched
 - [FIX] Fix scroll on save edit screen
+- [FIX] Update Infrared backend integration
 - [CI] Add https://github.com/LionZXY/detekt-decompose-rule
 - [CI] Enabling detekt module for android and kmp modules
 - [CI] Bump target SDK to 34